requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.1
python-dotenv>=1.0.0
Flask-WTF>=1.2.1
Pillow>=10.0.0
gunicorn>=21.2.0
reportlab>=4.0.7
openpyxl>=3.1.2

